Restore ability to handle errors in middleware
PR #21 added the ability to configure this library to throw errors, which is more canonical koa as it allows middleware to handle errors. Previously, the library would only emit an error on the `app` event emitter.

PR #31 removed this ability.

This restores the ability to use the more koa-like (and JavaScript-like) error-throwing behavior.